workaround bug where changing a file breaks analysis
Can reproduce by opening a file, hovering something, edit something, and
the hover is broken.

The root cause of the issue is the syntax tree requested in `hover` is
not the same one that the typechecker ran against so the pointer are all
wrong.

Worked around by returning the syntax tree the typechecker used in
`typechecker.Info` so we don't mismatch.

However, root cause is TBD