hack a working implementation of multi project graphqlrc files
The way this works is sort of a heuristic. It looks at the glob provided
for the graphql schemas and does a replacement s/.graphql/.gqlt and uses
that as a glob on the gqlt file to figure out which schema to use.

i.e.
```
projects:
    foo:
        schema: foo/**/*.graphql
```

We will then apply the glob `foo/**/*.gqlt` to the path of the `.gqlt` file we're
inspecting and do a reverse lookup and then determine which project (and
therefore schema) we should be using.

Tested with a particular example to ensure it at least kinda works.