Add a CLI `gqlt run <url> <query-file>`
`gqlt` still serves an LSP, I think it's convenient to keep one binary.

This contains a breaking change where `HTTPClient` is renamed to
`HTTPHandler`.

There is a more general `gqlt.Client` implementation defined in terms of
`http.RoundTripper`. The `Handler` and `Client` variants are trivially
defined in terms of that.